requirements.txt
openai
pytest
PyYAML
numpy
pandas
pyarrow